Add files via upload
The Solar System is a gravitationally bound system consisting of the Sun and the celestial bodies orbiting around it. It formed around 4.6 billion years ago from the collapse of a dense region in a molecular cloud, leading to the creation of the Sun and a protoplanetary disc. The system includes eight planets, numerous planetary satellites, asteroids, comets, and interstellar medium. The Solar System is located in the Orion Spur of the Milky Way Galaxy, a barred spiral galaxy that spans about 100,000 light years.